# backend/swagger.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/agentWallet": {
      "get": {
        "description": "Check if an agent wallet exists for the authenticated user and retrieve its status. Agent wallets are used to execute trades on Hyperliquid Exchange on behalf of users and are valid for 180 days with automatic 30-day rotations.",
        "operationId": "AgentWalletController_getAgentWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Agent wallet found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAgentWalletResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Agent wallet not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get agent wallet status",
        "tags": [
          "Agent Wallet"
        ]
      },
      "post": {
        "description": "Create a new agent wallet for the authenticated user. The wallet private key is securely stored and encrypted within Brother Terminal. After creation, the user must approve the wallet through Hyperliquid's agent wallet approval process.",
        "operationId": "AgentWalletController_createAgentWallet",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Agent wallet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAgentWalletResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new agent wallet",
        "tags": [
          "Agent Wallet"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate a user using EIP712 signature or API key and return access/refresh tokens",
        "operationId": "AuthController_authenticate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Authentication request with EIP712 signature or API key",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid input data"
          },
          "401": {
            "description": "Unauthorized - authentication failed"
          }
        },
        "summary": "Authenticate user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/eip712-message": {
      "get": {
        "description": "Generate EIP712 structured data for client-side signing",
        "operationId": "AuthController_getEIP712Message",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "query",
            "description": "User wallet address",
            "schema": {
              "example": "0x1234567890123456789012345678901234567890",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "EIP712 message generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEIP712MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid parameters"
          }
        },
        "summary": "Get EIP712 message to sign",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Use a valid refresh token to obtain a new access token and refresh token",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Refresh token request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refresh successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Invalidate refresh token and log out the user",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Logout request with refresh token",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid refresh token"
          }
        },
        "summary": "Logout user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api-keys": {
      "post": {
        "description": "Create a new API key for the authenticated user",
        "operationId": "ApiKeysController_createApiKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "API key creation request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Optional name for the API key",
                    "example": "Trading Bot Key"
                  }
                },
                "$ref": "#/components/schemas/Object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "apiKey": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create API key",
        "tags": [
          "API Keys"
        ]
      },
      "get": {
        "description": "Get all API keys for the authenticated user",
        "operationId": "ApiKeysController_getUserApiKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKeys": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string",
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "lastUsedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user API keys",
        "tags": [
          "API Keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brother Terminal Trading API",
    "description": "API for managing trading positions and orders on the Brother Terminal platform for Hyperliquid Engine.",
    "version": "3.0.0",
    "contact": {
      "name": "Brother Terminal",
      "url": "https://brother-terminal.xyz",
      "email": "dev@brother-terminal.xyz"
    }
  },
...
  "servers": [
    {
      "url": "https://api.brother-terminal.xyz",
      "description": "Production (Mainnet)"
    },
    {
      "url": "https://api-testnet.brother-terminal.xyz",
      "description": "Testnet"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "GetAgentWalletResponseDto": {
        "type": "object",
        "properties": {
          "agentWalletAddress": {
            "type": "string",
            "description": "Agent wallet address for Hyperliquid operations",
            "example": "0xabcdef1234567890abcdef1234567890abcdef12"
          },
          "agentName": {
            "type": "string",
            "description": "Agent identifier for this wallet",
            "example": "Brother Terminal API"
          },
          "status": {
            "type": "string",
            "description": "Wallet status - ACTIVE (ready), EXPIRED (needs renewal), or NOT_FOUND (needs creation)",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "NOT_FOUND"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "agentWalletAddress",
          "agentName",
          "status"
        ]
      },
      "CreateAgentWalletResponseDto": {
        "type": "object",
        "properties": {
          "agentWalletAddress": {
            "type": "string",
            "description": "Newly created agent wallet address - user must approve this on Hyperliquid",
            "example": "0xabcdef1234567890abcdef1234567890abcdef12"
          },
          "message": {
            "type": "string",
            "description": "Next steps instruction for user",
            "example": "Please approve this agent wallet on Hyperliquid Exchange to enable trading"
          }
        },
        "required": [
          "agentWalletAddress",
          "message"
        ]
      },
      "EIP712AuthDetailsDto": {
        "type": "object",
        "properties": {
          "signature": {
            "type": "string",
            "description": "EIP712 signature",
            "example": "0x1234567890abcdef..."
          },
          "timestamp": {
            "type": "number",
            "description": "Unix timestamp used in the signed message",
            "example": 1703872800
          }
        },
        "required": [
          "signature",
          "timestamp"
        ]
      },
      "ApiKeyAuthDetailsDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string",
            "description": "API key for authentication",
            "example": "abc123def456..."
          }
        },
        "required": [
          "apiKey"
        ]
      },
      "AuthenticateRequestDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "description": "Authentication method",
            "example": "eip712",
            "enum": [
              "eip712",
              "api_key"
            ]
          },
          "address": {
            "type": "string",
            "description": "User wallet address",
            "example": "0x1234567890123456789012345678901234567890",
            "pattern": "^0x[a-fA-F0-9]{40}$"
          },
          "details": {
            "description": "Authentication method specific details",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EIP712AuthDetailsDto"
              },
              {
                "$ref": "#/components/schemas/ApiKeyAuthDetailsDto"
              }
            ]
          }
        },
        "required": [
          "method",
          "address",
          "details"
        ]
      },
      "AuthenticateResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token for obtaining new access tokens",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "description": "Token type",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "description": "Token expiration time in seconds",
            "example": 900
          },
          "address": {
            "type": "string",
            "description": "User wallet address",
            "example": "0x1234567890123456789012345678901234567890"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "address"
        ]
      },
      "GetEIP712MessageResponseDto": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "object",
            "description": "EIP712 domain",
            "example": {
              "name": "Brother Terminal",
              "version": "1",
              "chainId": 1,
              "verifyingContract": "0x0000000000000000000000000000000000000000"
            }
          },
          "types": {
            "type": "object",
            "description": "EIP712 types",
            "example": {
              "Authentication": [
                {
                  "name": "address",
                  "type": "address"
                },
                {
                  "name": "timestamp",
                  "type": "uint256"
                },
                {
                  "name": "action",
                  "type": "string"
                }
              ]
            }
          },
          "primaryType": {
            "type": "string",
            "description": "Primary type for EIP712 signing",
            "example": "Authentication"
          },
          "message": {
            "type": "object",
            "description": "Message to sign",
            "example": {
              "address": "0x1234567890123456789012345678901234567890",
              "timestamp": 1703872800,
              "action": "authenticate"
            }
          },
          "timestamp": {
            "type": "number",
            "description": "Unix timestamp used in message",
            "example": 1703872800
          }
        },
        "required": [
          "domain",
          "types",
          "primaryType",
          "message",
          "timestamp"
        ]
      },
      "RefreshTokenRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "New JWT access token",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "description": "New refresh token",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tokenType": {
            "type": "string",
            "description": "Token type",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "number",
            "description": "Token expiration time in seconds",
            "example": 900
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn"
        ]
      },
      "LogoutRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token to invalidate",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Logout status message",
            "example": "Successfully logged out"
          }
        },
        "required": [
          "message"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}